# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: arca-lb Controller REST API
  description: REST API for managing VIPs and backends in arca-lb.
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/backends:
    get:
      tags: [Backends]
      summary: List backends for a VIP
      operationId: listBackends
      parameters:
        - name: vip_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackendListResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Backends]
      summary: Create backend
      operationId: createBackend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackendCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backend"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: VIP not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/backends/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Backends]
      summary: Get backend
      operationId: getBackend
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backend"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Backends]
      summary: Update backend
      operationId: updateBackend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackendUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backend"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Backends]
      summary: Delete backend
      operationId: deleteBackend
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    IPAddress:
      type: string
      description: IPv4 or IPv6 address.
...
    Backend:
      type: object
      properties:
        id:
          type: string
        vip_id:
          type: string
        ip:
          $ref: "#/components/schemas/IPAddress"
        weight:
          type: integer
          minimum: 1
          maximum: 100
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, vip_id, ip, weight, created_at, updated_at]
    BackendCreateRequest:
      type: object
      properties:
        vip_id:
          type: string
        ip:
          $ref: "#/components/schemas/IPAddress"
        weight:
          type: integer
          minimum: 1
          maximum: 100
      required: [vip_id, ip]
    BackendUpdateRequest:
      type: object
      properties:
        ip:
          $ref: "#/components/schemas/IPAddress"
        weight:
          type: integer
          minimum: 1
          maximum: 100
      additionalProperties: false
    BackendListResponse:
      type: object
      properties:
        backends:
          type: array
          items:
            $ref: "#/components/schemas/Backend"
        count:
          type: integer
      required: [backends, count]